Add files via upload
The bagana is a ten-stringed box-lyre of the Ethiopian Orthodox Tewahedo Church of
Ethiopia. Paraliturgical and solo instrument played only for
religious and meditative purposes, it often creates immediate and intense emotions for both players and listeners.
Based on informations collected during four fieldworks
held in Ethiopia (2002-2005), multidisciplinary analyses
have showed that these strong emotional reactions are created by sonorous and contextual means only. Structural and
timbral properties of the bagana repertoire are the most
salient ones playing a role in the arousal of the emotional
reactions, and the religious meaning and the values embodied in the instrument are the most important contextual
characteristics. These analyses show that bagana performances are meant create a specific time and space for emotion.